# frontend/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DR-OS Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/analysis/plans": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Analysis Plan",
        "operationId": "create_analysis_plan_v1_projects__project_id__analysis_plans_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisPlanRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/workflows": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_v1_projects__project_id__workflows_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_v1_projects__project_id__workflows_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/workflows/{workflow_instance_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_v1_projects__project_id__workflows__workflow_instance_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "workflow_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/workflows/{workflow_instance_id}/advance": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Advance Workflow",
        "operationId": "advance_workflow_v1_projects__project_id__workflows__workflow_instance_id__advance_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "workflow_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/workflows/{workflow_instance_id}/cancel": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Cancel Workflow",
        "operationId": "cancel_workflow_v1_projects__project_id__workflows__workflow_instance_id__cancel_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "workflow_instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/analysis-runs": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Analysis Run",
        "operationId": "create_analysis_run_v1_projects__project_id__analysis_runs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Analysis Runs",
        "operationId": "list_analysis_runs_v1_projects__project_id__analysis_runs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/analysis-runs/{run_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Analysis Run",
        "operationId": "get_analysis_run_v1_projects__project_id__analysis_runs__run_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvanceWorkflowRequest": {
        "properties": {
          "task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdvanceWorkflowRequest"
      },
...
      "AnalysisAgentResult": {
        "properties": {
          "template_id": {
            "type": "string",
            "minLength": 1,
            "title": "Template Id"
          },
          "template_version": {
            "type": "string",
            "minLength": 1,
            "title": "Template Version"
          },
          "parameter_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameter Json"
          },
          "preflight_checks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preflight Checks"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "template_id",
          "template_version"
        ],
        "title": "AnalysisAgentResult"
      },
      "AnalysisRunDetailResponse": {
        "properties": {
          "analysis_run": {
            "$ref": "#/components/schemas/AnalysisRunRead"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRead"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_run"
        ],
        "title": "AnalysisRunDetailResponse"
      },
      "AnalysisRunListResponse": {
        "properties": {
          "items": {
            "$ref": "#/components/schemas/Page_AnalysisRunRead_"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AnalysisRunListResponse"
      },
...
      "CancelWorkflowRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "CancelWorkflowRequest"
      },
...
      "CreateAnalysisPlanRequest": {
        "properties": {
          "study_goal": {
            "type": "string",
            "minLength": 10,
            "title": "Study Goal"
          },
          "dataset_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Dataset Ids"
          },
          "candidate_templates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Templates"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "study_goal",
          "dataset_ids"
        ],
        "title": "CreateAnalysisPlanRequest"
      },
      "CreateAnalysisPlanResponse": {
        "properties": {
          "workflow_instance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Instance Id"
          },
          "plan": {
            "$ref": "#/components/schemas/AnalysisAgentResult"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_instance_id",
          "plan"
        ],
        "title": "CreateAnalysisPlanResponse"
      },
      "CreateAnalysisRunRequest": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "template_id": {
            "type": "string",
            "format": "uuid",
            "title": "Template Id"
          },
          "params_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params Json"
          },
          "random_seed": {
            "type": "integer",
            "title": "Random Seed",
            "default": 0
          },
          "workflow_instance_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Instance Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id",
          "template_id"
        ],
        "title": "CreateAnalysisRunRequest"
      },
      "CreateAnalysisRunResponse": {
        "properties": {
          "analysis_run": {
            "$ref": "#/components/schemas/AnalysisRunRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_run"
        ],
        "title": "CreateAnalysisRunResponse"
      },
...
      "CreateWorkflowRequest": {
        "properties": {
          "workflow_type": {
            "type": "string",
            "minLength": 1,
            "title": "Workflow Type"
          },
          "runtime_backend": {
            "$ref": "#/components/schemas/WorkflowBackend",
            "default": "queue_workers"
          },
          "started_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "parent_workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Workflow Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_type"
        ],
        "title": "CreateWorkflowRequest"
      },
      "CreateWorkflowResponse": {
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/WorkflowInstanceRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow"
        ],
        "title": "CreateWorkflowResponse"
      },
...
      "GateEvaluationRead": {
        "properties": {
          "verification_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Id"
          },
          "gate_name": {
            "$ref": "#/components/schemas/GateName"
          },
          "target_kind": {
            "$ref": "#/components/schemas/LineageKind"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "status": {
            "$ref": "#/components/schemas/GateStatus"
          },
          "details_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details Json"
          },
          "evaluated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentName"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gate_name",
          "target_kind",
          "target_id",
          "status",
          "evaluated_at"
        ],
        "title": "GateEvaluationRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_WorkflowInstanceRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowInstanceRead"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "title": "Page[WorkflowInstanceRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowBackend": {
        "type": "string",
        "enum": [
          "queue_workers",
          "temporal"
        ],
        "title": "WorkflowBackend"
      },
      "WorkflowDetailResponse": {
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/WorkflowInstanceRead"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/WorkflowTaskRead"
            },
            "type": "array",
            "title": "Tasks"
          },
          "gate_evaluations": {
            "items": {
              "$ref": "#/components/schemas/GateEvaluationRead"
            },
            "type": "array",
            "title": "Gate Evaluations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow"
        ],
        "title": "WorkflowDetailResponse"
      },
      "WorkflowInstanceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "parent_workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Workflow Id"
          },
          "started_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "runtime_backend": {
            "$ref": "#/components/schemas/WorkflowBackend"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "workflow_type",
          "state",
          "runtime_backend",
          "started_at"
        ],
        "title": "WorkflowInstanceRead"
      },
      "WorkflowListResponse": {
        "properties": {
          "items": {
            "$ref": "#/components/schemas/Page_WorkflowInstanceRead_"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowState": {
        "type": "string",
        "enum": [
          "created",
          "retrieving",
          "retrieved",
          "structuring",
          "structured",
          "analyzing",
          "analyzed",
          "asserting",
          "asserted",
          "writing",
          "verifying",
          "approved",
          "blocked",
          "needs_human",
          "exported",
          "failed"
        ],
        "title": "WorkflowState"
      },
      "WorkflowTaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "workflow_instance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Instance Id"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "input_payload_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Payload Json"
          },
          "output_payload_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Payload Json"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "workflow_instance_id",
          "task_key",
          "task_type",
          "state",
          "input_payload_json",
          "output_payload_json",
          "retry_count",
          "created_at"
        ],
        "title": "WorkflowTaskRead"
      }
    }
  }
}